requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.42
phonenumberslite==8.13.42
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
